if db is empty, upgrade directly to the sqlalchemy model
We now have lots of versions. The initial install takes lots of time
especially for pg or sqlite on a slow disk

If we detect no versioning, we just use sqlaclhemy to create the tables from
the model
