+++ master/buildbot/db/migrate/versions/001_initial.py
@@ -18,6 +18,7 @@ from __future__ import with_statement
 import os
 import sqlalchemy as sa
 
+from buildbot.db.migrate_utils import test_unicode
 from buildbot.util import json
 from buildbot.util import pickle
 from twisted.persisted import styles
@@ -137,35 +138,6 @@ builds = sa.Table('builds', metadata,
                   sa.Column('start_time', sa.Integer, nullable=False),
                   sa.Column('finish_time', sa.Integer),
                   )
-
-
-def test_unicode(migrate_engine):
-    """Test that the database can handle inserting and selecting Unicode"""
-    # set up a subsidiary MetaData object to hold this temporary table
-    submeta = sa.MetaData()
-    submeta.bind = migrate_engine
-
-    test_unicode = sa.Table('test_unicode', submeta,
-                            sa.Column('u', sa.Unicode(length=100)),
-                            sa.Column('b', sa.LargeBinary),
-                            )
-    test_unicode.create()
-
-    # insert a unicode value in there
-    u = u"Frosty the \N{SNOWMAN}"
-    b = '\xff\xff\x00'
-    ins = test_unicode.insert().values(u=u, b=b)
-    migrate_engine.execute(ins)
-
-    # see if the data is intact
-    row = migrate_engine.execute(sa.select([test_unicode])).fetchall()[0]
-    assert isinstance(row['u'], unicode)
-    assert row['u'] == u
-    assert isinstance(row['b'], str)
-    assert row['b'] == b
-
-    # drop the test table
-    test_unicode.drop()
 
 
 def import_changes(migrate_engine):
