+++ master/buildbot/db/model.py
@@ -20,6 +20,8 @@ import sqlalchemy as sa
 from migrate import exceptions
 
 from buildbot.db import base
+from buildbot.db.migrate_utils import should_import_changes
+from buildbot.db.migrate_utils import test_unicode
 from buildbot.db.types.json import JsonObject
 from twisted.python import log
 from twisted.python import util
@@ -786,11 +788,21 @@ class Model(base.DBConnectorComponent):
                 # and, finally, upgrade using migrate
                 upgrade(engine)
 
-            # otherwise, this db is uncontrolled, so we just version control it
-            # and update it.
-            else:
+            # if we detect a pre-db installation, we run the migration engine
+            # to convert pickles
+            elif should_import_changes(engine):
                 version_control(engine)
                 upgrade(engine)
+            # otherwise, this db is new, so we dont bother using the migration engine
+            # and just create the tables, and put the version directly to latest
+            else:
+                # do some tests before getting started
+                test_unicode(engine)
+
+                Model.metadata.create_all(engine)
+                repo = migrate.versioning.repository.Repository(self.repo_path)
+
+                version_control(engine, repo.latest)
 
         # import the prerequisite classes for pickles
         check_sqlalchemy_migrate_version()
